Добавил блок с тестами. Сделал тесты сервисов и контроллеров моделей Pet.

+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/controllers/DogShelter/DogShelterPetController.java
@@ -12,7 +12,7 @@ import sky.pro.SkyDreamTeam.AnimalService.model.DogShelter.DogShelterPet;
 import sky.pro.SkyDreamTeam.AnimalService.service.DogShelter.DogShelterPetService;
 
 @RestController
-@RequestMapping(path ="dogShelter/Pet")
+@RequestMapping(path ="dogShelter/pet")
 public class DogShelterPetController {
     private final DogShelterPetService petService;
 
@@ -88,9 +88,17 @@ public class DogShelterPetController {
                     )
             )}
     )
-    @DeleteMapping("{id}")
-    public ResponseEntity deletePet(@PathVariable Long id) {
-        petService.deletePet(id);
+    @DeleteMapping
+    public ResponseEntity deletePet(@Parameter(description = "id животного")
+                                    @RequestParam(required = false) Long id,
+                                    @Parameter(description = "имя животного")
+                                    @RequestParam(required = false) String name) {
+        if (id != null) {
+            petService.deletePet(id);
+        }
+        if (name != null) {
+            petService.deletePet(name);
+        }
         return ResponseEntity.ok().build();
     }
 }
